+++ src/Output/Writers/EchoWriter.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 /**
  * ASCII Table
  *
- * @package   MarcinOrlowski\AsciiTable
+ * @package   MarcinOrlowski\TextTable
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
  * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
- * @link      https://github.com/MarcinOrlowski/php-ascii-table
+ * @link      https://github.com/MarcinOrlowski/php-text-table
  */
 
-namespace MarcinOrlowski\AsciiTable\Output\Writers;
+namespace MarcinOrlowski\TextTable\Output\Writers;
 
-use MarcinOrlowski\AsciiTable\Output\WriterContract;
+use MarcinOrlowski\TextTable\Output\WriterContract;
 
 class EchoWriter implements WriterContract
 {