+++ src/Renderers/DefaultRenderer.php
@@ -4,25 +4,25 @@ declare(strict_types=1);
 /**
  * ASCII Table
  *
- * @package   MarcinOrlowski\AsciiTable
+ * @package   MarcinOrlowski\TextTable
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
  * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
- * @link      https://github.com/MarcinOrlowski/php-ascii-table
+ * @link      https://github.com/MarcinOrlowski/php-text-table
  */
 
-namespace MarcinOrlowski\AsciiTable\Renderers;
+namespace MarcinOrlowski\TextTable\Renderers;
 
-use MarcinOrlowski\AsciiTable\Align;
-use MarcinOrlowski\AsciiTable\AsciiTable;
-use MarcinOrlowski\AsciiTable\Cell;
-use MarcinOrlowski\AsciiTable\Column;
-use MarcinOrlowski\AsciiTable\ColumnsContainer;
-use MarcinOrlowski\AsciiTable\Exceptions\ColumnKeyNotFoundException;
-use MarcinOrlowski\AsciiTable\Output\WriterContract;
-use MarcinOrlowski\AsciiTable\Row;
-use MarcinOrlowski\AsciiTable\Utils\StringUtils;
+use MarcinOrlowski\TextTable\Align;
+use MarcinOrlowski\TextTable\Cell;
+use MarcinOrlowski\TextTable\Column;
+use MarcinOrlowski\TextTable\ColumnsContainer;
+use MarcinOrlowski\TextTable\Exceptions\ColumnKeyNotFoundException;
+use MarcinOrlowski\TextTable\Output\WriterContract;
+use MarcinOrlowski\TextTable\Row;
+use MarcinOrlowski\TextTable\TextTable;
+use MarcinOrlowski\TextTable\Utils\StringUtils;
 
 class DefaultRenderer implements RendererContract
 {
@@ -31,7 +31,7 @@ class DefaultRenderer implements RendererContract
      *
      * @throws ColumnKeyNotFoundException
      */
-    public function render(AsciiTable $table, WriterContract $writer): void
+    public function render(TextTable $table, WriterContract $writer): void
     {
         $columns = $table->getColumns();
 
@@ -50,12 +50,10 @@ class DefaultRenderer implements RendererContract
             }
         } else {
             $label = 'NO DATA';
-            if (\mb_strlen($label) > $table->getTotalWidth()) {
-                $label = \mb_substr($label, 0, $table->getTotalWidth() - 1) . '…';
-            } else {
-                $label = StringUtils::pad($label, $table->getTotalWidth(), ' ', \STR_PAD_BOTH);
-
-            }
+            $tableTotalWidth = $this->getTableTotalWidth($table);
+            $label = (\mb_strlen($label) > $tableTotalWidth)
+                ? \mb_substr($label, 0, $tableTotalWidth - 1) . '…'
+                : StringUtils::pad($label, $tableTotalWidth, ' ', \STR_PAD_BOTH);
             $noData = \sprintf('| %s |' . PHP_EOL, $label);
             $writer->write($noData);
         }
@@ -294,6 +292,28 @@ class DefaultRenderer implements RendererContract
     {
         return $columns->getColumn($columnKey)->getTitleAlign();
     }
+
     /* ****************************************************************************************** */
+
+    /**
+     * Get total width of the visible table WITHOUT edges (so the full table width
+     * in output is then 2 (for left edge) + getTotalWidth() + 2 (for right edge).
+     */
+    protected function getTableTotalWidth(TextTable $table): int
+    {
+        $totalWidth = 0;
+
+        foreach ($table->getColumns() as $column) {
+            /** @var Column $column */
+            if ($column->isVisible()) {
+                $totalWidth += $column->getWidth();
+            }
+        }
+
+        // Next, we need to account column separators as well.
+        $totalWidth += ($table->getVisibleColumnCount() - 1) * \mb_strlen(self::HEADER_BORDER_CENTER);
+
+        return $totalWidth;
+    }
 
 }