+++ src/Traits/ArrayAccessTrait.php
@@ -4,17 +4,17 @@ declare(strict_types=1);
 /**
  * ASCII Table
  *
- * @package   MarcinOrlowski\AsciiTable
+ * @package   MarcinOrlowski\TextTable
  *
  * @author    Marcin Orlowski <mail (#) marcinOrlowski (.) com>
  * @copyright 2022 Marcin Orlowski
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
- * @link      https://github.com/MarcinOrlowski/php-ascii-table
+ * @link      https://github.com/MarcinOrlowski/php-text-table
  */
 
-namespace MarcinOrlowski\AsciiTable\Traits;
+namespace MarcinOrlowski\TextTable\Traits;
 
-use MarcinOrlowski\AsciiTable\ArrayableContract;
+use MarcinOrlowski\TextTable\ArrayableContract;
 use Traversable;
 
 /**