+++ CHANGES.md
@@ -6,9 +6,12 @@ Fast and flexible PHP library rendering plain ASCII tables.
 
 ## Changes
 
+* dev
+  * Main class name changed to `TextTable`. 
+
 * 1.7.0 (2022-10-19)
   * Corrected rendering tables with first/last column hidden.
-  * Attempt to render table with all colums hidden throws `NoVisibleColumnsException`.
+  * Attempt to render table with all columns hidden throws `NoVisibleColumnsException`.
   * All custom exception class names now ends with `Exception`.